+++ client/examples/get_funds.rs
@@ -20,7 +20,6 @@ async fn main() -> Result<()> {
     // Create a client instance
     let client = Client::builder()
         .with_node(&node_url)? // Insert the node here
-        .with_node_sync_disabled()
         .finish()?;
 
     let secret_manager = SecretManager::Mnemonic(MnemonicSecretManager::try_from_mnemonic(