+++ client/examples/ledger_nano.rs
@@ -21,7 +21,6 @@ async fn main() -> Result<()> {
     // Create a client instance
     let client = Client::builder()
         .with_node(&node_url)? // Insert your node URL here
-        .with_node_sync_disabled()
         .finish()?;
 
     let ledger_nano = LedgerSecretManager::new(false);