+++ client/examples/node_api_core/05_post_block_raw.rs
@@ -20,10 +20,7 @@ async fn main() -> Result<()> {
     });
 
     // Create a client with that node.
-    let client = Client::builder()
-        .with_node(&node_url)?
-        .with_node_sync_disabled()
-        .finish()?;
+    let client = Client::builder().with_node(&node_url)?.finish()?;
 
     let min_pow_score = client.get_min_pow_score().await?;